Refactor SequentialIndexer based on review
changelog_begin
changelog_end
